Simplify build and platform-specific stuff.
- Put all platform specific code into `platform.h' and `platform.c'.
- Unify Makefiles.
- Remove subdirectories that contained platform-specific build and code.
