--- freebsd/util.h
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2010 Marco Peereboom <marco@peereboom.us>
- *
- * Permission to use, copy, modify, and distribute this software for any
- * purpose with or without fee is hereby granted, provided that the above
- * copyright notice and this permission notice appear in all copies.
- *
- * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
- * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
- * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
- * ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
- * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
- * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
- * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
- */
-/*
- * This file tries to make up for the difference between OpenBSD's
- * <util.h> and FreeBSD's <libutil.h>.
- */
-
-/*
- * fmt_scaled(3) specific flags. (from OpenBSD util.h)
- */
-
-#define FMT_SCALED_STRSIZE	7	/* minus sign, 4 digits, suffix, null byte */
-
-int	fmt_scaled(long long number, char *result);
