+++ xombrero.h
@@ -34,8 +34,6 @@
 
 #include <sys/types.h>
 #if defined(__linux__)
-#include "linux/util.h"
-#include "linux/tree.h"
 #include <bsd/stdlib.h>
 #include <X11/Xlib.h>
 #include <sys/un.h>
@@ -44,7 +42,6 @@ void		arc4random_buf(void *, size_t);
 u_int32_t	arc4random_uniform(u_int32_t);
 # endif
 #elif defined(__APPLE__)
-#include "osx/util.h"
 #include "osx/tree.h"
 #include <stdlib.h>
 #include <X11/Xlib.h>
@@ -53,8 +50,6 @@ u_int32_t	arc4random_uniform(u_int32_t);
 #include <sys/ucred.h>
 #define LOGIN_NAME_MAX MAXLOGNAME
 #elif defined(__FreeBSD__)
-#include <libutil.h>
-#include "freebsd/util.h"
 #include <sys/tree.h>
 #include <X11/Xlib.h>
 #include <sys/un.h>
@@ -63,20 +58,16 @@ u_int32_t	arc4random_uniform(u_int32_t);
 #elif defined(__DragonFly__)
 #include <sys/param.h>
 #include <sys/un.h>
-#include <libutil.h>
-#include "dragonfly/util.h"
 #include <sys/tree.h>
 #include <X11/Xlib.h>
 #define LOGIN_NAME_MAX MAXLOGNAME
 #elif defined(__NetBSD__)
 #include <sys/param.h>
 #include <sys/un.h>
-#include "netbsd/util.h"
 #include <sys/tree.h>
 #include <X11/Xlib.h>
 #define LOGIN_NAME_MAX_MAXLOGNAME
 #elif defined(__OpenBSD__)
-#include <util.h>
 #include <sys/tree.h>
 #include <X11/Xlib.h>
 #include <sys/un.h>
@@ -90,6 +81,8 @@ uint32_t	arc4random_uniform(uint32_t);
 #include <sys/socket.h>
 #include <sys/stat.h>
 #include <sys/time.h>
+
+#include "platform.h"
 
 #include <gtk/gtk.h>
 #include <gdk/gdk.h>
